Backend Address, Checkout, Customer, Order m.m

+++ src/main/java/com/disc/ecommerce/entity/Product.java
@@ -30,7 +30,7 @@ public class Product {
     @Column(name = "description")
     private String description;
     @Column(name = "unit_price")
-    private BigDecimal unitPrice;
+    private Long unitPrice;
     @Column(name = "image_url")
     private String imageUrl;
